Initial attempt at logic to create a topic. Does not yet use tenant name.
