+++ src/ServiceFabricPubSub/RequestRouter/Controllers/RequestController.cs
@@ -8,51 +8,48 @@ namespace RequestRouterService.Controllers
     [RequestAuthorization]
     public class RequestController : ApiController
     {
+        private const string TenantApplicationAdminServiceName = "Admin";
+
         // PUT api/tenantId/topicName
         public async Task<HttpResponseMessage> Put(string tenantId, string topicName)
         {
-            HttpResponseMessage adminServiceResponseMessage;
             HttpResponseMessage responseMessage = new HttpResponseMessage(HttpStatusCode.InternalServerError);
-
-            // TODO: Check if the tenant is valid. (TenantApplication/tenantId)
-            HttpServiceUriBuilder adminServiceUriBuilder = new HttpServiceUriBuilder()
-            {
-                ServiceName = $"TenantApplication/{tenantId}",
-                ServicePathAndQuery = ""
-            };
-
-            using (HttpClient httpClient = new HttpClient())
-            {
-                adminServiceResponseMessage = await httpClient.GetAsync(adminServiceUriBuilder.Build());
-            }
 
             // Assuming the message body will contain the content for the data to put to the topic.
             string messageBody = await this.Request.Content.ReadAsStringAsync();
 
+            ReverseProxyPortResolver portResolver = new ReverseProxyPortResolver();
+            int portNumber = await portResolver.GetReverseProxyPortAsync();
+
             HttpServiceUriBuilder builder = new HttpServiceUriBuilder()
             {
-                ServiceName = "TenantApplication/TopicSerivce"
+                // TODO: Need to use the tenant name instead of 'TenantApplication'.
+
+                // TODO: Get port number dynamically
+                PortNumber = 19081,
+                //ServiceName = $"{tenantName}/{TenantApplicationAdminServiceName}/api/topics/topicName"
+                ServiceName = $"TenantApplication/{TenantApplicationAdminServiceName}/api/topics/{topicName}"
             };
 
             // Call the Topic Service to post the message.
-            HttpResponseMessage response;
+            HttpResponseMessage topicResponseMessage;
             using (HttpClient httpClient = new HttpClient())
             {
-                 response = await httpClient.GetAsync(builder.Build());
+                topicResponseMessage = await httpClient.PutAsync(builder.Build(), null);
             }
 
-            if (response != null && response.IsSuccessStatusCode)
+            if (topicResponseMessage != null && topicResponseMessage.IsSuccessStatusCode)
             {
                 responseMessage.StatusCode = HttpStatusCode.Accepted;
 
-                var msg = await response.Content.ReadAsStringAsync();
+                var msg = await topicResponseMessage.Content.ReadAsStringAsync();
 
                 // TODO: do something with the response.
             }
             else
             {
                 responseMessage.StatusCode = HttpStatusCode.InternalServerError;
-                responseMessage.ReasonPhrase = response?.ReasonPhrase ?? "Internal error";
+                responseMessage.ReasonPhrase = topicResponseMessage?.ReasonPhrase ?? "Internal error";
             }
 
             return responseMessage;
